hybrid_multimodal_stack/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hybrid-multimodal-stack"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.1",
  "transformers>=4.40",
  "accelerate>=0.30",
  "pyyaml>=6.0",
  "numpy>=1.24",
  "Pillow>=10.0",
  "huggingface_hub>=0.23",
  "safetensors>=0.4"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
